refactor(deploy): Dynamically associate sidecars
